backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI + LangGraph backend with MCP integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.4",
    "langchain-mcp-adapters>=0.1.0",
    "python-dotenv>=1.2.1",
    "httpx>=0.28.0",
    "sse-starlette>=2.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "mcp>=1.0.0",
]

[dependency-groups]
dev = []
